arena_camera: fix for image conversion to pointcloud

+++ catkin_ws/src/arena_camera/include/arena_camera/arena_camera_parameter.h
@@ -287,7 +287,7 @@ public:
   bool auto_flash_line_3_;
 
   /* ADDITIONAL PARAMETERS */
-  int amplitude_gain_;
+  double amplitude_gain_;
   std::string exposure_time_selector_;
   std::string conversion_gain_;
   std::string operating_mode_;
